Android App-ePrescription
An Android app which digitizes the entire process of Medical
Prescriptions. The patient can consult his doctor from the comforts of
his home via messaging and video calling. The doctor can prescribe
suitable drugs for the patient remotely. This prescription also serves
as a medical history of the patient which can be accessed at a touch of
the finger.